refactor: 역할 별로 패키지 분리 및 라인 정리
구현 기능 목록에 명시된 대로 역할별로 패키지 분리 (Domain, Util, Exception)

코드 라인 정리
- IntelliJ fomatter 적용 (naver-intellij-formatter.xml)
