+++ src/main/java/racingcar/Application.java
@@ -1,7 +1,9 @@
 package racingcar;
 
-import static racingcar.Input.*;
-import static racingcar.Output.*;
+import static racingcar.utils.Input.*;
+import static racingcar.utils.Output.*;
+
+import racingcar.domain.Car;
 
 public class Application {
 	public static void main(String[] args) {
@@ -10,13 +12,14 @@ public class Application {
 		Car[] cars = inputCarNameToCarObj(carNames);
 
 		int coin = inputCoinNumber();
+
 		while (coin-- > 0) {
 			inputRandomNumberToCar(cars);
 			outputCarPosition(cars);
 		}
+
 		outputRaceResult(cars);
 	}
-
 
 }
