+++ src/main/java/racingcar/domain/Judge.java
@@ -1,6 +1,6 @@
-package racingcar;
+package racingcar.domain;
 
-import static racingcar.Car.*;
+import static racingcar.domain.Car.*;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,18 +9,18 @@ import org.assertj.core.util.Lists;
 
 public class Judge {
 
-	public static List<String> judging(Car[] cars){
+	public static List<String> judging(Car[] cars) {
 		List<Car> carList = Lists.list(cars);
 		sortingByPosition(carList);
 
-		return getCarPositionPerRace(carList);
+		return getRaceWinnerList(carList);
 	}
 
-	private static List<String> getCarPositionPerRace(List<Car> carList) {
+	private static List<String> getRaceWinnerList(List<Car> carList) {
 		int max = carList.get(0).getPosition();
 
 		List<String> winnerList = new ArrayList<>();
- 		for (Car car : carList) {
+		for (Car car : carList) {
 			if (max == car.getPosition()) {
 				winnerList.add(car.getName());
 			}
