+++ src/main/java/racingcar/exception/CarNameNotValidException.java
@@ -1,10 +1,9 @@
 package racingcar.exception;
 
-
-public class CarNameNotValidException extends IllegalArgumentException{
+public class CarNameNotValidException extends IllegalArgumentException {
 	private static final String MESSAGE = "[ERROR] 이름은 5자 이하만 가능하다.";
 
-	public CarNameNotValidException(){
+	public CarNameNotValidException() {
 		System.out.println(MESSAGE);
 	}
 }
