+++ src/main/java/racingcar/exception/NumberFormatNotValidException.java
@@ -1,9 +1,9 @@
 package racingcar.exception;
 
-public class NumberFormatNotValidException extends IllegalArgumentException{
+public class NumberFormatNotValidException extends IllegalArgumentException {
 	private static final String MESSAGE = "[ERROR] 시도 횟수는 숫자여야 한다.";
 
-	public NumberFormatNotValidException(){
+	public NumberFormatNotValidException() {
 		System.out.println(MESSAGE);
 	}
 }
