+++ src/main/java/racingcar/utils/Output.java
@@ -1,10 +1,10 @@
-package racingcar;
+package racingcar.utils;
 
-import static racingcar.Car.*;
 
 import java.util.List;
 
-import org.assertj.core.util.Lists;
+import racingcar.domain.Car;
+import racingcar.domain.Judge;
 
 public class Output {
 	private static final String RESULT_MESSAGE = "실행 결과";
@@ -15,32 +15,31 @@ public class Output {
 	}
 
 	public static void outputRaceResult(Car[] cars) {
-
-		StringBuilder sb = new StringBuilder();
 		System.out.println(RESULT_MESSAGE);
-
-		printWinnerList(cars, sb);
-		System.out.println(WINNER_MESSAGE + trimResult(sb));
+		printWinnerList(cars);
 	}
 
-	private static void printWinnerList(Car[] cars, StringBuilder sb) {
+	private static void printWinnerList(Car[] cars) {
+		StringBuilder sb = new StringBuilder();
+
 		List<String> ouputList = Judge.judging(cars);
-		for(String name : ouputList){
-			sb.append(name+", ");
+		for (String name : ouputList) {
+			sb.append(name + ", ");
 		}
+
+		System.out.println(WINNER_MESSAGE + trimWinnerResultString(sb));
 	}
 
 	private static void printCarStatus(Car[] cars) {
 		StringBuilder sb = new StringBuilder();
+
 		for (Car car : cars) {
-			sb.append(car.toString()+"\n");
+			sb.append(car.toString() + "\n");
 		}
 		System.out.println(sb.toString());
 	}
 
-
-
-	private static String trimResult(StringBuilder sb) {
+	private static String trimWinnerResultString(StringBuilder sb) {
 		String result = sb.toString();
 		result = result.substring(0, result.length() - 2);
 		return result;
